add file upload

+++ controllers/admin.js
@@ -1,4 +1,5 @@
 const Product = require('../models/product');
+const fileHelper = require('../util/file');
 const { validationResult } = require('express-validator');
 
 ////////////////////
@@ -18,7 +19,24 @@ exports.getAddProduct = (req, res, next) => {
 // POST ADD PRODUCT
 exports.postAddProduct = (req, res, next) => {
   console.log('Products.js', req.body);
-  const { title, price, imageUrl, description } = req.body;
+  const { title, price, description } = req.body;
+  const image = req.file;
+  console.log(image);
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title,
+        price,
+        description
+      },
+      errorMessage: 'Attached file is not an image.',
+      validationErrors: []
+    });
+  }
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
@@ -29,13 +47,15 @@ exports.postAddProduct = (req, res, next) => {
       product: {
         title,
         price,
-        imageUrl,
         description
       },
       errorMessage: errors.array()[0].msg,
       validationErrors: errors.array()
     });
   }
+
+  const imageUrl = image.path;
+
   const product = new Product({
     title,
     price,
@@ -46,7 +66,6 @@ exports.postAddProduct = (req, res, next) => {
   product
     .save()
     .then(result => {
-      // console.log(result);
       console.log('PRODUCT CREATED!');
       res.redirect('/admin/products');
     })
@@ -92,7 +111,9 @@ exports.getEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   // console.log('THIS IS THE BODY: ', req.body);
-  const { title, price, imageUrl, description } = req.body;
+  const { title, price, description } = req.body;
+  const image = req.file;
+  console.log('IMAGE:', image);
   const errors = validationResult(req);
 
   if (!errors.isEmpty()) {
@@ -104,7 +125,6 @@ exports.postEditProduct = (req, res, next) => {
       product: {
         title,
         price,
-        imageUrl,
         description,
         _id: prodId
       },
@@ -118,11 +138,17 @@ exports.postEditProduct = (req, res, next) => {
       if (product.userId.toString() !== req.user._id.toString()) {
         return res.redirect('/');
       }
+
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+        console.log('UPDATED IMG:', product.imageUrl);
+      }
       return product
         .updateOne({
           title,
           price,
-          imageUrl,
+          image,
           description
         })
         .then(result => {
@@ -131,6 +157,7 @@ exports.postEditProduct = (req, res, next) => {
         });
     })
     .catch(err => {
+      console.log(err);
       const error = new Error(err);
       error.httpStatusCode = 500;
       return next(error);
@@ -160,7 +187,15 @@ exports.getProducts = (req, res, next) => {
 // POST DELETE PRODUCT
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not found.'));
+      }
+      // console.log(product.imageUrl);
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log('PRODUCT DELETED!');
       res.redirect('/admin/products');
